requirements.txt
cryptography
pip
attrs
argparse
pyOpenSSL
redis
cffi
setuptools
six
koji
python-dateutil
requests
requests-gssapi
defusedxml
pika
Twisted
suds
MarkupSafe
Werkzeug
click
the-new-hotness
Jinja2
blinker
itsdangerous
PyJWT
gssapi
decorator
pyasn1
Automat
certifi
crochet
wrapt
python-bugzilla
termine
incremental
hyperlink
constantly
urllib3
greenlet
idna
charset-normalizer
pycparser
Flask
referencing
jsonschema
jsonschema-specifications
SQLAlchemy
Flask-JWT-Extended
flask-cors
django-email-users
python-dotenv
typing_extensions
fedora_messaging
anitya_schema
pyasn1_modules